project/requirements.txt
gradio
Pillow
facenet-pytorch
torch
opencv-python
grad-cam